fix: `update_type` of the tag packet is a bit mask
The client currently always sends both updates at once (`0b00000011` = `3`), so the time was never updated correctly.

I also introduced the unknown type now, in case the clients sends values bigger than 3 that are unknown to the server.
(The update then might be neither for time nor for `is_it`.)