fix: synchronization issues
- Send empty `TagPacket` and `CapturePacket` on new connections, to reset old data back that other players might still have in their puppet from an
earlier connection.
- Cache and send `CostumePacket`, `CapturePacket`, `GamePacket` and `PlayerPacket` to (re-)connecting players.

Same change in the C# server:
- https://github.com/Sanae6/SmoOnlineServer/pull/49

---

The last `TagPacket` is not cached, because it could contain an update type that isn't complete.
Instead a packet is created from `time` and `is_seeking` for (re-)connecting players.

Same change in the C# server:
- https://github.com/Sanae6/SmoOnlineServer/pull/52

---

As this server deletes the player data on disconnects, the cached packets from earlier re-connected connections are no longer available.

Therefore the saved packets don't need to be cleared for fresh connections (after a restart).

---

But this can cause sync issues for new players now, if ealier re-connecting players use a mod version below 1.3.0.
Because the mod below 1.3.0 does not send all of the important game packets again after a reconnect.

I'm ignoring this for now, because version 1.4.0 of the mod is required to connect to a UDP server that has a fix for this.

But if this rust server software is used for TCP-only a client with an older mod version might connect and run into these issues.
For now I've added two TODO comments in `client.rs` and `coordninator.rs` to mark where a change is needed.