fix: support for general game mode packets
There are compatibility issues between different game modes (H&S, Sardines, Freeze-Tag) regarding the existing packets.
Therefore I'm defining that the first 4 bits of the existing u8 update_type indicate the game mode from now on.
- add: parse, remember and display the `game_mode` for the JSON API

The existing TagUpdate parsing had an issue that all undefined update_types map to Unknown, which were broadcast as 0 instead of the actual value from
the client.
- fix: parse and broadcast all possible values for `update_type` in other game modes

Additionally always parsing `is_it` as a bool had an issue that if the byte is not a bool it'd modify values higher than 1 to 1.
- fix: parse `is_it` as a bool only if `game_mode` and `update_type` are known for it

Lastly extracting the time only makes sense if that's what the bytes are used for.
- fix: extract `seconds` & `minutes` only if `game_mode` and `update_type` are known for it

---

These changes keep parsing the time and tagged status for Hide & Seek and Sardines as before but not for other game modes anymore.
Packets that aren't parsed in that way (for other game modes) are now distributed 1:1 to other clients instead of parsing it wronlgy and therefore
modifying them which might break compatibility.

E.g. Freeze-Tag currently uses game mode packets with update_types (1, 2, 4, 8).
They overlap with those of H&S and Sardines in an incompatible way (1, 2, 3).
Without a game mode indicator packets w/ update_type 1 or 2 could be wrongly parsed by the client (or server).
(E.g. when multiple players are connected and are in different game modes.)

Also the data layout for all of the Freeze-Tag packets is different.
Therefore there is no `is_it`, `seconds` or `minutes` that can be extracted from these packets in a meaningful way.
(`is_runner` could be interpreted as `it_it` though, but doesn't really convey the same meaning.)
This would definitely be an issue for `update_type` 2 (ROUNDSTART).
Because the second byte in the packet isn't a bool (`is_it`) but a u8 (`round_time`) instead.
Parsing it as a bool by the server and then broadcasting it to other players breaks compatibility.
(Because values higher than 1 would be modified to be 1.)